[entropy.client] Fix accidental deletion of preserved libs when still in use.
Entropy preserved libs handling code did not check all the followed symlink
before removing them. This causes the deletion of unexpected libraries during the
removal of packages. For example, app-emulation/vmware-workstation contains
the following symlink:

(A) /opt/vmware/lib/vmware/lib/libssl.so.1.0.1/libssl.so.1.0.1 -> (B) /usr/lib64/libssl.so.1.0.1

Upon removal of vmware-workstation, A is garbage collected and deleted,
if it's a symlink, the link is followed and (B) is removed as well, without
validating the action (i.e. checking if a package is still owning (B).)

This commit fixes this. See Sabayon bug #5812.

+++ lib/entropy/client/interfaces/package/preservedlibs.py
@@ -285,6 +285,19 @@ class PreservedLibraries(object):
         for lib_path in self._follow(library_path):
             root_lib_path = self._root + lib_path
 
+            # Guard against cross package symlinking. Example:
+            # vmware-workstation pkg conaining symlinks to libssl.so.x
+            # See Sabayon bug #5182.
+            lib_path_pkg_ids = self._inst_repo.isFileAvailable(
+                lib_path, get_id=True)
+            if lib_path_pkg_ids:
+                pkg_atoms = [self._inst_repo.retrieveAtom(x) for x in lib_path_pkg_ids]
+                failed.append(
+                    (root_lib_path,
+                     '%s owned by %s' % (lib_path, ', '.join(pkg_atoms)))
+                )
+                continue
+
             try:
                 os.remove(root_lib_path)
             except (OSError, IOError) as err:
